Add support for default makefile parameters
This allows to successfully build all configuration on default even if some of
them require to disable some programmers to not fail, e.g. because a
library was not installed yet or is not available on the system at all.

These default parameters are only applied if the user does not supply any.